% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{get_genes_by_cluster}
\alias{get_genes_by_cluster}
\title{Get genes by cluster}
\usage{
get_genes_by_cluster(se, target = c("genes", "samples"))
}
\arguments{
\item{se}{SummarizedExperiment or ExpressionSet object}

\item{target}{"genes" (clustering must be done on genes)}
}
\value{
List of gene vectors, one per cluster
}
\description{
Get genes by cluster
}
\examples{
library(SummarizedExperiment)
mat <- matrix(rnorm(40), nrow = 10)
rownames(mat) <- paste0("gene", 1:10)
colnames(mat) <- paste0("sample", 1:4)
se <- SummarizedExperiment(assays = list(counts = mat))
SummarizedExperiment::rowData(se)$Cluster <- rep(1:2, each = 5)
get_genes_by_cluster(se, target = "genes")
}
