% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{perform_cluster_go_enrichment}
\alias{perform_cluster_go_enrichment}
\title{Perform GO enrichment analysis per gene cluster}
\usage{
perform_cluster_go_enrichment(
  genes_by_cluster,
  orgdb,
  ont = "BP",
  keyType = "SYMBOL"
)
}
\arguments{
\item{genes_by_cluster}{List of gene vectors (output of get_genes_by_cluster)}

\item{orgdb}{OrgDb object (e.g., org.Hs.eg.db)}

\item{ont}{Ontology for GO: "BP", "MF", "CC" (default "BP")}

\item{keyType}{Character; key type used in OrgDb (default = "SYMBOL")}
}
\value{
List of enrichGO results per cluster
}
\description{
Perform GO enrichment analysis per gene cluster
}
\examples{
library(org.Hs.eg.db)
clusters <- list(
  Cluster_1 = c("TP53", "EGFR", "BRCA1"),
  Cluster_2 = c("MTOR", "AKT1", "MAPK1")
)
results <- perform_cluster_go_enrichment(clusters, org.Hs.eg.db, ont = "BP")
}
