% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_plots.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Plot dendrogram of hierarchical clustering}
\usage{
plot_dendrogram(hc, k = NULL)
}
\arguments{
\item{hc}{hclust object}

\item{k}{Optional, cut tree into k clusters}
}
\value{
Dendrogram plot
}
\description{
Plot dendrogram of hierarchical clustering
}
\examples{
mat <- matrix(rnorm(25), 5, 5)
hc <- hclust(dist(mat))
plot_dendrogram(hc, k = 2)
}
