% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_expr_matrix}
\alias{set_expr_matrix}
\title{Replce expression matrix in object}
\usage{
set_expr_matrix(se, exprs_mat, assay_name = "counts")
}
\arguments{
\item{se}{A SummarizedExperiment or ExpressionSet object.}

\item{exprs_mat}{A numeric matrix to replace the expression data.}

\item{assay_name}{Name of assay to replace.}
}
\value{
Updated object of the same type as input.
}
\description{
Replce expression matrix in object
}
\examples{
mat <- matrix(rpois(12, lambda = 10), nrow = 3)
rownames(mat) <- paste0("gene", 1:3)
colnames(mat) <- paste0("sample", 1:4)
se <- SummarizedExperiment::SummarizedExperiment(
    assays = list(counts = mat)
)
new_mat <- mat + 1
se <- set_expr_matrix(se, new_mat, assay_name = "counts")
get_expr_matrix(se, "counts")
}
